ai can add company description on url based searching
