+++ types/types.ts
@@ -3,6 +3,7 @@ export interface Company {
   name: string
   slug: string
   description: string
+  companyType: string
 }
